Add files via upload
The project is developed using html, css and javascript for frontend development
Xampp server and mysql for the database storing
php is user for the both frontend and server connection

The main of the project finding the chef according the specific category, charges, chefnames. They are many categories like chinnese,juices,indian,tandoori..etc. After the searching the chefname, category then the chef details will be displayed. The displayed details are chefname, charges, category,availability..etc, after clicking the booknow button then the details will be stored in database and booking cofimation page will be displayed.